+++ docs/changelog.rst
@@ -9,6 +9,7 @@ CHANGELOG
 
 - New ``LEIParser`` to import touristic content and event from LEI touristic data system
 - New ``XMLParser`` to import content from XML
+- ApidaeTrekParser: import trek's contact info into description
 
 **Bug fixes**